backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiosqlite>=0.19.0

# Security libraries
slowapi>=0.1.9
secure>=0.3.0
pathvalidate>=3.2.0  # Filename and path validation/sanitization
python-magic>=0.4.27  # Optional: better file type detection
filetype>=1.2.0  # Optional: lightweight alternative

# Validation
pydantic>=2.5.0

# Async file I/O
aiofiles>=23.0.0

# ML/AI dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
accelerate>=0.20.0  # Batch size optimization and distributed training
Pillow>=10.0.0
numpy>=1.24.0

# System monitoring (optional - no longer used)
# psutil>=5.9.0

# Development tools (optional)
bandit>=1.7.5  # Security linting
safety>=2.3.5  # Dependency vulnerability scanning

